NetBSD support: revise Log()
different from other systems, NetBSD's tv_usec in struct timeval is
still 32bit (not 64bit).

sprintf()'s format uses %lu(64bit unsigned) for milliseconds and
this will not work proerly when 32bit value is feeded.

normally gettimeofday() returns tv_usec value less than 1,000,000,
so formatting with %u(32bit unsigned) and casting the value to
unsigned int may not be harm (but looks ugly...).

+++ Log.cpp
@@ -119,7 +119,7 @@ void Log(unsigned int level, const char* fmt, ...)
 
 	struct tm* tm = ::gmtime(&now.tv_sec);
 
-	::sprintf(buffer, "%c: %04d-%02d-%02d %02d:%02d:%02d.%03lu ", LEVELS[level], tm->tm_year + 1900, tm->tm_mon + 1, tm->tm_mday, tm->tm_hour, tm->tm_min, tm->tm_sec, now.tv_usec / 1000U);
+	::sprintf(buffer, "%c: %04d-%02d-%02d %02d:%02d:%02d.%03u ", LEVELS[level], tm->tm_year + 1900, tm->tm_mon + 1, tm->tm_mday, tm->tm_hour, tm->tm_min, tm->tm_sec, (unsigned int)now.tv_usec / 1000U);
 #endif
 
 	va_list vl;
